requirements.txt
python-dotenv
praw
imgurpython
Pillow
